requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
email-validator==2.0.0
python-dotenv==1.0.0
passlib==1.7.4
Flask-SocketIO==5.3.4
Flask-Session==0.8.0
gevent==23.9.1
gevent-websocket==0.10.1
requests==2.31.0
python-socketio==5.8.0
PyNaCl==1.5.0
Flask-talisman==1.1.0
bcrypt==4.3.0
